Add picture element (#69)
I have (for the most part) been successful in whitelisting attributes to specific tags in my app's initializer, but new syntax (if you can call a tag that was shipped in Chrome in 2014 "new" 😄 ) has been harder. Finally I realised that the initializer wasn't enough, so here's a PR adding the picture tag. More (possibly "turbo-frame") will potentially follow. 

The grey type that indicates you're in a picture block is ever so slightly darker than the rest, I think, but it's using the correct font and font-size, so I hope it's good enough.